# bot/docs/swagger.yaml
basePath: /
definitions:
  models.APIResponse:
    properties:
      data: {}
      message:
        example: Passwords retrieved successfully
        type: string
      reason:
        type: string
      request_url:
        example: /password/get_password
        type: string
      timestamp:
        example: "2025-01-17T20:58:06Z"
        type: string
    type: object
  models.Password:
    properties:
      password:
        type: string
      site:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
  description: This is an API for managing user passwords
  title: Password Management API
  version: "1.0"
paths:
  /password/get_password:
    get:
      consumes:
      - application/json
      description: Get passwords by site name for the user
      parameters:
      - description: User ID
        in: query
        name: userID
        required: true
        type: string
      - description: Site name
        in: query
        name: site
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get Passwords By Site
      tags:
      - Password
  /password/get_password__userID/{userID}:
    get:
      consumes:
      - application/json
      description: Get all passwords for a user by user ID
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get All Passwords
      tags:
      - Password
  /password/post_password:
    post:
      consumes:
      - application/json
      description: Create a new password for the user
      parameters:
      - description: Password
        in: body
        name: Password
        required: true
        schema:
          $ref: '#/definitions/models.Password'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Create Password
      tags:
      - Password
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"